- Modify TOs to allow that all parameters are optional, so they can be {@code null} when not used: set Object types instead of primitive types. - In EntityTO, No more checks if the ID is null since TO may be retrieved without ID. - Modify test according new TOs and add some tests. - Fix MySQLRelationDAOIT, didn't use TOComparator in shouldGetAnatEntityRelations(). - Javadoc edits - Remove unused ORDER BY.

+++ bgee-dao-api/src/test/java/org/bgee/model/dao/api/expressiondata/CallTOTest.java
@@ -438,6 +438,7 @@ public class CallTOTest extends TestAncestor {
         
         int count = 10;
         callTO.setMinConditionCount(count);
-        assertEquals("Incorrect minConditionCount set/get", count, callTO.getMinConditionCount());
+        assertEquals("Incorrect minConditionCount set/get", 
+                count, (int) callTO.getMinConditionCount());
     }
 }
